DateIndex alias for int
... and move code to `dates.py`.
